chore(vmm): organized
